Move regression test to its own file
